bluez5-init: reset bluetooth qualcomm chip before initialization
The power regulator connected to the qca6554 chip is always on, which causes
the Bluetooth part to not work correctly after a software-reset.
This commit asserts momentarily the BT_EN line during the start-up sequence
to reset the Bluetooth controller so that it is in a predictable state after a
reset.

https://jira.digi.com/browse/DEL-2623

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>