install_linux_fw_sd: expand variables when setting bootcmd to avoid redefining them
The setting of 'bootcmd' in the script was done using single quotes, which
doesn't expand variables. As a consequence the following variables must
be defined again (during the execution of the second part of the script):
 - mmcdev
 - INSTALL_LINUX_FILENAME
 - INSTALL_ROOTFS_FILENAME

This patch changes the single quotes with double quotes, so that these
variables are expanded during the setting of 'bootcmd' with the values
assigned at the begining of the installation script (notice these are
not dynamically generated so there is no risk to expand them).

At the same time we need to escape with a backslash:
 - double quotes containing strings
 - variables that we don't want to expand (like the return value $?)

Signed-off-by: Hector Palacios <hector.palacios@digi.com>