u-boot: disable TF initramfs for CC6UL
At the moment there is no support for rootfs encryption for the CC6UL,
so there is not a ramdisk in the boot image. But with the initial
addition of TF support, the u-boot boot script was being on-the-fly
updated for TF regardless of the platform, making the CC6UL unable to
boot when TF was enabled.

This commit fixes the problem, by just changing the u-boot boot script
when TF is enabled only for the CC6.

https://jira.digi.com/browse/DEL-2754

Signed-off-by: Javier Viguera <javier.viguera@digi.com>