kernel-module-qualcomm: add support to override the region code
The region code is programmed in the OTP bits. We want to allow to be able to
override this behavior by setting the new value as a property in the device
tree called "regulatory-domain".
This can be done by setting the variable "regdomain=<code>" in uboot or well
by defining that entry in the device tree.

https://jira.digi.com/browse/DEL-2799

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>