meta-digi-dey: trustfence: Do not disable console access by default.
While performing usability testing on the TrustFence documentation, it has
been noted that in order to follow the secure boot instructions the
console needs to be enabled.

We have now moved the secure console section to the end of the
documentation so that disabling the console is the last configuration to
make in a secure system.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>