ccimx6ul: suspend: unload wireless module always before suspending
The suspend script was unloading the wireless module driver *only* if the
interface 'wlan0' was previously up. If the interface is down the module
does not unload and the system fails to suspend.

This patch removes the module always if it appears in the list of loaded
modules.

While on it, move the unloading of wireless after the unloading of
Bluetooth, to be in line with the reversed order in which they are loaded
on the resume_interfaces() call.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3241