meta-digi-arm: linux-dey: create postfunc for trustfence
The kernel recipe was modifying the device tree blobs in place within the
kernel build temporal directory. This can cause problems after several
compilations, only the deployed artifacts should be signed/encrypted.

The deployment of the DTBs is done by do_deploy_appends in other layers which
are appended after this recipe, so it is required to use a postfunc to do the
trustfence related process after the deployment of all the artifacts.

https://jira.digi.com/browse/DEL-3388

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>