swupdate: use version 2016.10
There is new functionality in that version that we need for our firmware
update solution, so append the git-based recipe and set the revision to
the 2016.10 version.

Also provide a customized build configuration file. Notice that the
defconfig file is actually a full '.config'. It needs to be a full
config file because otherwise the anonymous python function in the
recipe is not able to gather all the build time dependences.

https://jira.digi.com/browse/DEL-3355

Signed-off-by: Javier Viguera <javier.viguera@digi.com>