meta-digi-dey: pulseaudio: load HDMI sink on hotplug
When hotplugging the HDMI cable after booting the target, pulseaudio doesn't
load the HDMI sink. This commit fixes this by adding a udev rule and a script
that prompts pulseaudio to detect the HDMI sound card and load its sink, as
well as making it the default one.

If alsa-utils is available, a faulty command is run to load the card without
any audio output. Otherwise, module-udev-detect is unloaded and reloaded.

This is based on commit fca507d316ded1e3126ba6db265b7fe871550fae.

https://jira.digi.com/browse/DEL-2327

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>