meta-digi-arm: firmware-qualcomm: Update board calibration files.
These have been provided by the embedded hardware team and they are to
be used for certification testing. The calibration corresponds to the
internal antenna on the module. If using the current external antenna on the
CC6UL SBC they work fine, although they could be tweaked to improve.

The BDF files provided were generated from the original ones, by running a
test tree for output power calibration for QCA6564A-3 chip. This test
tree is called
‘FCCSP_Rome3.x_WLAN_CLPC_cal_and_WLAN_BT_Verify_Y8611-H341.xtt’ and it
belongs to the Qualcomm Design Package ‘DP25-WL420-100’.
The test tree is to be run in the QSPR software tool, part of the QDART
software tools set provided by Qualcomm.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>