meta-digi-dey: pulseaudio: reduce pulseaudio loading time
On boot, pulseaudio looks for the sound cards' corresponding profile sets, and
uses a default configuration if none is found. This was causing pulseaudio to
take up to 5 seconds trying every possible default profile.

This commit fixes this by creating a specific profile set for each of the sound
cards on the ccimx6sbc so pulseaudio doesn't spend time trying different
profiles that don't apply to our platform. It also adds udev rules so
pulseaudio knows which configuration to load for each card.

https://jira.digi.com/browse/DEL-2218

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>