ccimx6sbc: pulseaudio: change module-suspend-on-idle timeout to 0
In pulseaudio, module-suspend-on-idle controls how much time a sink spends
in the 'idle' state before going to 'suspended'. This means the ALSA device
associated with the sink stays busy for a few seconds after loading or using
the sink.

This commit adds the timeout=0 parameter when loading module-suspend-on-idle.
Now the sink will go to the 'suspended' state right after being loaded or used.

https://jira.digi.com/browse/DEL-2218

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>