ccimx6sbc: pulseaudio: Make pulseaudio load the sgtl5000 sink
The profiles created for the sgtl5000 sound card didn't have any priorities
assigned, making pulseaudio map a different profile than before and causing the
sink to remain unloaded.

This commit fixes this by assigning a priority to each profile. The values
were taken directly from the default profile to assure that pulseaudio's
behavior is the same using one profile or the other.

https://jira.digi.com/browse/DEL-3602

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>