meta-digi-arm: fix encrypted env support when using ubootenv
The caam encryption check was only done from fw_env_write and fw_env_read
functions, which are not called when using the functions exported as a library.

Move the check_caam_encryption() call to fw_env_open(), which is called from
all code paths. A similar check for AES encryption cannot be moved because it
requires the AES key as an argument.

https://jira.digi.com/browse/DEL-3616

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>