install_linux: fix failure to boot into recovery
On a CC6, the 'env default -a' command deletes the 'soc_family' variable
that is used on the U-Boot boot scripts to configure the name of the
device tree file to load.

Then if we try to boot into recovery directly (without a reset) because
the variable is not available, it fails to compose the correct name for
the fdt, and gives the error:

  reading uImage--ccimx6sbc-id129.dtb
  ** Unable to read file uImage--ccimx6sbc-id129.dtb **

Force a reset, so the 'soc_family' variable is available again and the
correct fdt_file is loaded.

While on it, add some blank lines to the 'recovery reboot' message, so
it's easier to read.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>