kernel-module-qualcomm: fix issue with _scan_callback at module unload
Protect the invocation of the _scan_done() callback function with the
global lock to avoid that it is called while the module is being unloaded
and the data structures have been freed.

Additionally, the commit also adds a patch to reduce the log level of the
driver that is logging some annoying messages.

https://jira.digi.com/browse/DEL-3607
https://jira.digi.com/browse/DEL-3393

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>