packagegroup-dey-gstreamer: fix override for CC6UL
As of DEY-2.0 "mx6" was an override value for CC6, but not for the
CC6UL, so we used it to restrict the packages to be installed only in
the CC6.

Now in DEY-2.2 "mx6" is also a correct override value for the 6UL
family, so correct the override to prevent those gstreamer packages to
be installed in 6UL machines.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>