meta-digi-dey: recipes-connectivity: bluez5: Allow to override baudrate
Reviewing the patches while updating the QCA firmware to a new release,
I encountered this patch. On the original PR it says:

"Do not use hcitattach to reconfigure the baudrate but set it directly in the
Bluetooth FW file; this save us from some syncronism problems if we are not
using HW flow control."

Current spins of the module already use hw flow control. Also, not allowing
hciattach to override the baudrate does not seem right to me. It would be
enough with passing a 3M baudrate to hciattach to match the firmware
default.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>