meta-digi-dey: fix bluetooth interface initialization
With the upgrade to Morty, there is a new bootscript that is provided
by Poky. This bootscript is just launching the generic bluetooth daemon,
but it's not doing any hardware initialization.

The fix is to split the hardware initialization to a different
bootscript 'bluetooth-init' and launch it also on boot.

https://jira.digi.com/browse/DEL-3855

Signed-off-by: Javier Viguera <javier.viguera@digi.com>