meta-digi-dey: disable ICU support for ccimx6ul
ICU support was not included in the rootfs in previous DEY versions
but it's included now through runtime dependences of 'harfbuzz' and
'libical'.

The support is not strictly needed, as ICU is a library for Unicode
support for applications:

http://site.icu-project.org/home

This commit disables ICU support directly for 'harfbuzz' and adds a
patch for 'libical' that allows to disable it. Notice that it's only
disabled for the CC6UL, as the CC6 has no rootfs space problems.

Only the 'libicudata' library is 25MB:

25M  /usr/lib/libicudata.so.57.1

https://jira.digi.com/browse/DEL-3854

Signed-off-by: Javier Viguera <javier.viguera@digi.com>