init-ifupdown: remove wrong sed command
This 'sed' command was used in the past, when DHCP or MANUAL was
configured in the '/etc/network/interfaces' (for cellular). This is no
longer true as we always configure the interface with a custom (manual)
script.

This was breaking the rest of interfaces when configured to use 'dhcp':

iface eth0 inet manual
       umanualc_opts -S -b >/dev/null &

Signed-off-by: Javier Viguera <javier.viguera@digi.com>