pulseaudio: disable memfd for CC6
Pulseaudio 9.0 relies on the syscall memfd_create, which was introduced
in kernel 3.17, but the version currently supported for the CC6 SBC is
3.14.

There is a corner case where the configure script does not detect the
actual kernel version, thus enabling memfd_create. When Pulseadio tries
to call memfd_create, it shows error messages, so disable memfd
explicitly.

https://jira.digi.com/browse/DEL-3923

Signed-off-by: Alejandro Navarro <alejandro.navarro@digi.com>