meta-digi-dey: add aws iot sdk embedded-c recipe
The AWS IoT device SDK for embedded C is a collection of C source files that
can be used in embedded applications to securely connect to the AWS IoT
platform.

The SDK code is available in GitHub and includes the library code and several
sample applications.

A user must provide from its 'local.conf':
  * AWS_IOT_CERTS_DIR: Absolute path to the directories where the required
    certifies are stored:
      * Root CA.
      * Device signed certificate.
      * Device private key.

And optionally, values for the rest of variables defined in the 'awsiotsdk-c'
bbclass.

This recipe is building the library code and the binaries for the samples:
  * shadow_sample
  * shadow_sample_console_echo
  * subscribe_publish_sample

See https://github.com/aws/aws-iot-device-sdk-embedded-C.

https://jira.digi.com/browse/DEL-4101

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>