modemmanager: update to fix QMI bug and add Ublox patches for XBee
QMI modems didn't work with the version of
modemmanager that the dey-2.2-r1 release was
using.

We were using two patches for uBlox plugin to use XBee Cellular 3G
that are now merged in modemmanager master.

This updates modemmanager to use the commit where QMI
problem was fixed and also include uBlox patches.

https://jira.digi.com/browse/DEL-4192

Signed-off-by: Salvador Penalva <salvador.penalva@digi.com>