u-boot-fw-utils: dynamically create fw_env.config for NAND-based platforms
A mechanism was introduced in U-Boot to dynamically locate the environment
copies within a NAND partition, skipping bad blocks and making full use of
the partition blocks.
To trigger this mechanism in Linux tools (fw_printenv/fw_setenv) the
configuration file /etc/fw_env.config must have both copies entries
pointing to the same offset.

This commit substitutes the hardcoded configuration file for CC6UL with a
stub and generates one on an RPM post-script basing on:

- Supplied DEY variables (with default values per platform):
  - UBOOT_ENV_OFFSET: starting offset of the environment within the
    partition.
  - UBOOT_ENV_SIZE: size of each copy of the environment.
  - UBOOT_ENV_RANGE: size in the partition that the environment copies may
    occupy, starting from UBOOT_ENV_OFFSET. If undefined, the whole
    partition is used.

- Calculated values:
  - NAND Erase block size
  - Number of blocks available for the environment copies

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-4256