kernel-module-qualcomm: fix two udev rules managing qcom wireless driver
Udev comes with a bunch of generic rules for the hardware. One of them:

/lib/udev/rules.d/80-drivers.rules

is able to load the qualcomm wireless module.

At the same time our new rule is needed, so it prepares the firmware
before the module is loaded.

Both rules running collide with following error message:

udevd[181]: '/etc/udev/scripts/qualcomm.sh' [184] terminated by signal 1 (Hangup)

So to fix this, we rename the specific rule so it's parsed later, and
then we use the hard assignation '=' instead of '+=' to override the
values that come from the previous parsed rule.

Also clean-up a bit the 'qualcomm.sh' script and the modprobe file,
given that now we know the standard drivers rule won't run for this
interface.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>