suspend: adapt script for NetworkManager integration
Now the suspend script brings down all the wireless interfaces that are
managed by ifupdown command (in the '/etc/network/interfaces' file) and
then unloads the wireless driver.

On resume the script triggers an event so udev reloads the wireless
driver and then brings up the interfaces that were brought down on
suspend.

https://jira.digi.com/browse/DEL-4288

Signed-off-by: Javier Viguera <javier.viguera@digi.com>