networkmanager: check the PPP connection
These values configure the PPP daemon to check that the connection is
still alive (using an echo) and kill itself otherwise. On that event,
NetworkManager takes over and tries to restore the connection.

This is specially useful for modems with only one serial port, because
they are not able to signal the connectivity loss to the host (the only
data channel is used for the network data itself).

https://jira.digi.com/browse/DEL-4653
https://jira.digi.com/browse/DEL-4650

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>