imx6ul: ubifs: increase max leb count for linux
The current value only allows partitions of up to 8 MiB, which is not
enough for the 0x04 variant (which uses a linux partition of 24 MiB by
default).

This new value allows for up to 32 MiB (assuming 128 KiB erase block size).

Also add comments explaining the -c value and maximum partition sizes
supported.

https://jira.digi.com/browse/DEL-4971

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>