hostapd: generate unique SSID for SoftAP
If two or more targets are in the same environment, all of them will
broadcast the same SSID, which may be a problem when a client wants
to connect to a specific device.

Use the last bytes of the non-virtual wireless MAC for the name of the
SSID, a random value with no hexadecimal characters is used if the MAC
is not available.
Also drop 'wpa2aes' from the SSID.

https://jira.digi.com/browse/DEL-4129

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>