curl: enable 'ares' build time option
This is required to use '--dns-interface' runtime parameter to instruct
'curl' to use a specific interface for DNS resolution.

We need this in the context of the network failover support, where
NetworkManager does the connectivity check, by trying to connect to a
configured URL through the different interfaces available, and for this
check it uses 'libcurl' underneath.

If you try to use this functionality without this build time option
enabled, you get:

curl: (4) A requested feature, protocol or option was not found built-in
in this libcurl due to a build-time decision.

https://jira.digi.com/browse/DEL-4787

Signed-off-by: Javier Viguera <javier.viguera@digi.com>