networkmanager: connectivity check: set interface for DNS resolution
Use the same interface that it's being checked for the DNS resolution.

This makes connectivity check work in an scenario where the primary
interface has lost connectivity (for example ethernet) but you still
have connectivity through a secondary interface (e.g. wireless).

What happens in that case is that even though you have correct
connectivity in the secondary interface the check fails because it does
not use that secondary interface to resolve the name of the test URL.
The result is that the connectivity check assumes that this secondary
interface is also failing and it penalizes the metrics in the routing
table.

This commit uses libcurl's CURLOPT_DNS_INTERFACE to set the DNS
interface. Notice that this requires 'libcurl' to be compiled with
'--enable-ares'.

https://jira.digi.com/browse/DEL-4787

Signed-off-by: Javier Viguera <javier.viguera@digi.com>