modemmanager: split udev rules for XBee Cellular
The XBee Cellular modem is expected to be at ttymxc1 on CC6UL and
at ttymxc4 on CC6. The changes are needed because in v4.9.11 kernel
the /sys/devices entry for CC6UL and CC6 are the same for the TTYs,
causing that on CC6UL ModemManager was also searching for modems on
ttymxc4.

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>
https://jira.digi.com/browse/DEL-5163