bluez5: fix out-of-bounds access in SDP server (CVE-2017-1000250)
Backport upstream commit from poky layer to be applied to bluez5 v5.33

Original message:
All versions of the SDP server in BlueZ 5.46 and earlier are vulnerable to an
information disclosure vulnerability which allows remote attackers to obtain
sensitive information from the bluetoothd process memory. This vulnerability
lies in the processing of SDP search attribute requests.

https://jira.digi.com/browse/DEL-5287

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>