packagegroup-dey-gstreamer: define installed packages per machine
This allows to specify a different set of gstreamer packages to be
installed per machine. It's mainly used for the CC6UL to define a
minimum set of packages that allows to play a WEBM video, and at the
same time save some space as this platform does not have much storage.

The space saved with this change for the CC6UL and the default QT image
is about ~5MB.

https://jira.digi.com/browse/DEL-4987

Signed-off-by: Javier Viguera <javier.viguera@digi.com>