AWS Greengrass: set the GG Core daemon relative path in the init script
The recipe modifies the Greengrass Core init script to properly configure the
daemon relative path. This daemon location is different depending on the
Greengrass version being used.

This commits also moves the patch for Greengrass Core 1.0.0 to the
'greengrass-1.0.0' directory in preparation for adding Greengrass Core 1.1.0
recipe in the next commit.

https://jira.digi.com/browse/DEL-5368

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>