trustfence-cst: avoid warnings about cst tarball
When parsing the recipe, a warning is shown because the tarball is only found
in the downloads folder. However this is expected as it cannot be distributed.

As a workaround, add the tarball to the SRC_URI variable only when Trustfence
is active. That way the warning is not shown in all other cases.

This was incorrectly removed in commit 14fc51147f6e.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>