libsoc: patch library to allow udev events to complete
After exporting a GPIO or PWM, we need to give some time for udev rules
to complete, before actually trying to access the newly created entries
in the sysfs.

This solves a problem, where udev was setting the mode and group of
the newly created files, so they are accessible for users (not root)
belonging to that group.

For traceability and to be sure the patch applies without conflicts,
this commit also sets a fixed revision to use from the libsoc
repository.

https://jira.digi.com/browse/DEL-5389

Signed-off-by: Javier Viguera <javier.viguera@digi.com>