bluez5.46: remove testing flag when starting the service
In bluez5.41 we needed to specify the "--experimental" flag when starting the
bluetooth service in order to use some specific features (see d434043447a93).
In bluez5.46 (in fact starting in bluez5.44) code was reorganized and now that
compilation option is called "--testing" but does not exist as flag, because
some of the "experimental" features has been mode to standard features, so
there is not need of indicating any extra flag when starting the service.

https://jira.digi.com/browse/DEL-5624

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>