bluez5.46: remove testing flag
There is not need to specify the testing flag for bluez5 compilation. That flag
is new starting at bluez5.44, and in the past it was called experimental. In
the past it made sense to pass that flag to compile specific services, but most
of them has been migrated to standard features, and experimental has been
renamed to testing, but we do not need any of those testing features.

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>