u-boot-fw-utils: fix environment access functions
In the old u-boot-fw-utils 2016.x the function 'fw_env_close' wrote
the environment to the media. That is no longer the case, now that
function only frees the memory used, and 'fw_env_flush' is used
to write the environment.

This was causing that userspace application using this library were not
able to write changes to the environment.

https://jira.digi.com/browse/DEL-5644

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>