trustfence-sign-tools: add missing dependency
util-linux is needed as the scripts use 'hexdump'. In previous versions
of Yocto the tool from the host was used, but this has been fixed in
Yocto 2.3

Also fix the other dependencies: the -native suffix is not needed nor
correct, as the native and nativesdk classes already map packages in
the DEPENDS variable to the corresponding packages.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>