build.sh: remove references to 'manufacturing.xml' file
The file 'manufacturing.xml' file in the manifests repository does not exist.
When building manufacturing artifacts, we must use the 'manufacturing/morty'
branch, that already contains the right repositories to clone.

For a manufacturing job, this commit only adds the 'meta-digi-mfg' if it is not
already there.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>
(cherry picked from commit 379ca469dff949e09c667d1a448012881b8812d2)
Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>