firmware-qualcomm: add support for QCA6574 in ConnectCore 8X
The firmware-qualcomm recipe now provides the Bluetooh and WiFi firmware
for QCA6564 (CC6UL, CC6PLUS) and QCA6574 (CC8X).

The Bluetooth firmware is the same for both chip models.

Regarding the Wifi firmware:
- If using the QCA6564, the proprietary firmware is used.
- If using the QCA6574, the recipe allows to select between the proprietary
  or the community driver (via QUALCOMM_WIFI_DRIVER), and will install the
  associated firmware. For the community driver, the firmware files have
  been copied from linux-firmware repo at SHA1 7518922bd5b9, except for
  the file 'board.bin' which needs to be substituted with the proprietary
  one by Qualcomm. Default, is to use the community driver.

The do_install() has been improved for re-using the variables with the list
of files.

The package has been renamed basing on the QCA_MODEL variable which can
be set per platform.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>