ccimx8x: change kernel to Image.gz and support all SOM variants in DT
U-Boot's 'dboot' command now supports uncompressing Image.gz.
This will be the default kernel image type for the CC8X.

This also changes the names of the device tree artifacts, that must be
updated on the boot script. While on it, add support for all the SOM
variants (wireless and non-wireless).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>