image_types: remove potential sub-folder from device tree files
The boot.(ext) image class was looping over ${KERNEL_DEVICETREE}
variable which is normally the list of device tree filenames.
The loop however does not take into account the possibility that
this list of files are within a sub-folder, as it happens with
ARM64 device tree files in the kernel tree.

Use 'basename' to remove any potential subfolder on the iterations.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6060