ccimx6/6qp: imx-gst1.0-plugin: add missing symbol to fix compilation
This package depends on libg2d, and the latest version of this library's
sources for imx6 platforms are currently unavailable. Among other things, the
newer sources add a symbol needed by imx-gst1.0-plugin to compile, so this
patch manually adds the symbol as a macro in the package's source code.

https://jira.digi.com/browse/DEL-6078

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>