imx-mkimage: update recipe to use the 8qxp_beta2 branch
In order for Yocto to append all the necessary firmware binaries to the U-Boot
image, this recipe needs to be up to date.

Also, remove the older hardcoded SRCREV in our .bbappend file. The file that
gets patched hasn't been modified since beta1, so our patch apllies cleanly to
beta2.

https://jira.digi.com/browse/DEL-6085

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>