imx-boot: add support for the imx-boot package
This package is the combination of U-Boot and the other i.MX firmware binaries.
In our case, these binaries are: the SCFW, the ATF and the DCD.

The recipe has been ported from the meta-fsl-bsp-release repo, with a couple of
important changes:

    * Support for our platform (based on the mx8qxpa0, which uses A0-specific
      parameters for mkimage and appends the DCD explicitly)
    * Support for multiple UBOOT_CONFIG values (NXP only uses one UBOOT_CONFIG
      string per platform; so the compile, install and deploy functions had to
      be tweaked to build both 1GB and 2GB variants for our platform).

Our DCD patch used in imx-mkimage is also used, since both packages share the
same repo and imx-boot is the one responsible for generating the final,
bootable image.

https://jira.digi.com/browse/DEL-6085

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>