meta-digi-dey: xserver-xf86-config: Use fbdev device for mainline
The meta-freescale layer assumes that when you define the use-mainline-bsp
the platform is going to use the etnaviv gpu driver. This is not the case
for the CC6UL as the mxsfb.c driver is still not DRM capable.

Probably this will change in v4.19 as there is a DRM capable mxsfb.c
driver that will make it possible to use the armada xf86 driver with the
etnaviv fbdev emulation.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>