build-github: fix construction of variables from platforms with dashes
Commit d981f999f0afc245002f7a81cf445350854afece introduced a new
platform with dashes on its name, which broke Jenkins build because
middle variables are being created using the name of the platform
and dashes are not allowed in bash variable names.

To fix it, this commit changes the dashes to underscores on the fly.

This is ported from a2092450faaa6bc332d08a56cf3e49b97755dc3c.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>