meta-digi-dey: bluez5: Adapt bluetooth-init to mainline use
The Digi vendor kernel uses a dedicated HCI UART driver with support for
In-Band Sleep (IBS). This driver, when used with a QCA vendor ROME plugin
for hciattach, is able to upload a new firmware version to the Bluetooth
chipset, as well as set the MAC address directly on the firmware.

Mainline BSP uses the HCI H4 driver instead and the "qualcomm" hciattach
plugin which is not able to upload firmware to the QCA6564 chipset.

As such, when using mainline the chipset uses the ROM firmware and the
bluetooth init script has been adapted accordingly.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>