imx-boot, imx-sc-firmware: adapt recipes for step B0 of the i.MX8X CPU
These adaptations include changing the mkimage targets and removing any
handling of the DCD files, since they are now included in the SCFW. Note that
these changes break support for step A0 of the silicon.

The md5sum of the imx-sc-firmware tarball will be updated once the step B0 CPUs
are ready.

https://jira.digi.com/browse/DEL-6158

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>