linux-4.18.y: Avoid compilation error
This error only appears when the build host does not have the openssl
development libraries (libssl-dev) installed. If they are, the build
process will use them.

The error is:

|
/home/builder/jenkins/workspace/dey-rocko-use-mainline-bsp/projects/ccimx6ulsbc/tmp/work-shared/ccimx6ulsbc/kernel-source/scripts/extract-cert.c:21:25:
fatal error: openssl/bio.h: No such file or directory
| compilation terminated.
| scripts/Makefile.host:90: recipe for target 'scripts/extract-cert'
failed
| make[3]: *** [scripts/extract-cert] Error 1
| make[3]: *** Waiting for unfinished jobs....

This commit makes openssl-native a dependency and passes the include
path to the build.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>