imx-boot: make do_populate_lic depend on packages dependant
When the imx-boot recipe is added in EXTRA_IMAGEDEPENDS the
license file is being referenced in image postcommand
write_deploy_manifest, but the licenses of the dependant
packages are missing. This leads to the license files of the
recipes to be missing when write_deploy_manifest runs,
causing build errors to appear.

https://jira.digi.com/browse/DEL-6251

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>